avformat/vapoursynth: Port to API v4, add audio
Port from the v3 API to the v4 API which has audio support.

Additionally, support multiple outputs to take advantage of the audio
support. Since there doesn't seem to be a way to get a list of outputs,
we just probe them. Dirty, but it works.

Also add a log hander just in case it might be useful.

Add some options some users might want to be able to tweak.
The max script file size has been upgraded to 16 MiB because that is
what VSScript evaluateFile() function does. To waste less memory, it'll be
shrunk to their minimal size.